Fixes rest timeout.
Former-commit-id: fa5fbc05d3099bfb85e32829e26f2905831fd290

+++ src/main/java/ca/qc/ircm/genefinder/rest/RestConfiguration.java
@@ -9,5 +9,5 @@ public interface RestConfiguration {
    *
    * @return timeout for REST requests
    */
-  public long timeout();
+  public int timeout();
 }
